changed state from an interface to an abstract class and included the field vars, refactored code to make use of this change.
Changed AnimationSequence.java to Camera.java, waits for the physics simulation to reach stability, halts it, takes a "snapshot", and then resumes it.
  -called by paint, which still runs concurrently with the physics thread.

Wrote an actuator handler, which works like a priority queue, but uses a CycleQueue to do the main physics.
  -can, and probably should be given the functionality to decide where actuators are
   placed in the cycle.

Wrote a new type of actuator ActuationCommand.java which is completely specified as an object by its constructor. next commit will include javadocs

Added the CycleQueue class, exactly what it sounds like.

rewrote Demo0.java to use new features, and the actuator sequence.