issue 2 implemented, use spacebar to switch between cameras.
current issues:

There are certain conditions which cause character-pill collision to be triggered more than once.
-if the character and a pill start out collided, multiple collision events trigger.
-also this very occasionally has happened while in play.

collision() is called much less frequently than simpleUpdate(), especially during times which are, for whatever reason, computationally intense. since repositioning after wall collision happens in collision(), the movement can appear choppy.

simpleUpdate does not need to be called as often as it is. I have not been able to find a way to slow it down though.

I would be curious to know how often collision is being called. I had to crop the walls to prevent them from colliding with each other, but I suspect that they are still close enough to put some strain on the collision DETECTION code, which happens in the background. it's possible that there is room for optimization there by specifying each of the walls in a seperate collision group.